дз4

+++ repository/index.js
@@ -1,17 +1,26 @@
 const Contact = require('../model/contact')
 
-const listContacts = async () => {
-  const results = await Contact.find({})
+const listContacts = async (userId) => {
+  const results = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription createdAt updatedAt'
+  })
+  // const {limit = 2, offset = 0} = query
+  // const searchOptions = { owner: userId }
+  // const results = await Contact.paginate(searchOptions, { limit, offset })
   return results
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findById(contactId)
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription createdAt updatedAt'
+  })
   return result
 }
 
 const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId })
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: userId })
   return result
 }
 const addContact = async (body) => {
@@ -19,9 +28,9 @@ const addContact = async (body) => {
   return result
 }
 
-const updateContact = async (contactId, body) => {
-  const result = await Contact.findByIdAndUpdate(
-    contactId,
+const updateContact = async (contactId, body, userId) => {
+  const result = await Contact.findOneAndUpdate(
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true },
   )
